+++ lib/SpeechWrap.cpp
@@ -16,7 +16,7 @@ void SpeechWrap::reset()
     cache_size = 0;
 }
 
-void SpeechWrap::load(int16_t *din, int len)
+void SpeechWrap::load(float *din, int len)
 {
     in = din;
     in_size = len;
@@ -35,7 +35,7 @@ void SpeechWrap::update(int offset)
     memcpy(cache, in + in_offset, cache_size * sizeof(int16_t));
 }
 
-int16_t &SpeechWrap::operator[](int i)
+float &SpeechWrap::operator[](int i)
 {
     return i < cache_size ? cache[i] : in[i - cache_size];
 }
